pyproject.toml
[dependency-groups]
dev = [
    "fastapi>=0.133.1",
    "python-dotenv>=1.2.1",
    "shopifyapi>=12.7.0",
    "uvicorn>=0.41.0",
]
